+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 1.3.0
+
+### Added:
+
+* Added the ability to set an __events__ property which is set to the EventManager class
+
+```typescript
+import { EventManager, EventListener, INextFn } from '@thenja/event-manager';
+
+export class UserService {
+  events = new EventManager();
+
+  userSignIn() {
+    this.events.emit('user-sign-in', {});
+  }
+}
+```
+
 # 1.2.0
 
 ### Added: